Part 1 - Use a 2d hashmap to store cross compartment wrappers.
Now we use a flat hashmap to store CCWs, which when we want to nuke wrappers
to a target compartment, we need to iterate every single one to find them. But if we
use a 2d hashmap to store CCWs by their target compartment, we can walk through
only the wrappers that we're targeting.